setup routes with express.js
this file is regarding the setup of routes with express.js in the node.js where we use middleware in order to do that. we install the cors which help us in restricting the CORS error in the server and we use custom middleware for most parts of the 404 error and we use the builtin middleware to handle the url encoded data.
this file also includes the program for the css and 404 error.