MVC model with CRUD-REST API
in this folders we have programmed a MVC model which consist of crud and rest api. the rest api look after the server that have been designed using the node.js and express.js ; whereas the crud api resides in the rest api which help to manipulate the data . we used the thunder client api to check if the API programmed can be worked or not.