user password authenticastion
this folder consists of creating user password authentcation using hash and salts to our node.js and express.js programs. this program also mention about how to add user registration and authentication routes. we use the package of bcrypt in this program which help us with the hash and salt