mongoDB initialisation
this folder consists of all the files which are for the initialisation of the mongoDB. we have downloaded the mongoose package installed in the package.json and  set up a MongoDB database and connect to it with Mongoose from your Node.js application. MongoDB uses a NoSQL database structure with collections and documents.