mongoDB schemas & data models
this folder consists of the files that are related to the MongoDB schemas and the data models that we created a database in the MongoDB atlas with a new cluster. this file also includes the supporting repositories and documents which i used to create a database in the MongoDB using the mongoose application.